# feat(persistence): Comprehensive Security, Performance & Testing Overhaul