--- a/ruv-swarm/crates/ruv-swarm-persistence/src/tests/mod.rs
+++ b/ruv-swarm/crates/ruv-swarm-persistence/src/tests/mod.rs
@@ -4,3 +4,7 @@ mod crud_tests;
 mod query_tests;
 mod storage_tests;
 mod transaction_tests;
+mod property_tests;
+mod security_tests;
+mod concurrent_tests;
+mod test_utils;